// WYNKWeb/src/app/Models/DoctorMaster.model.ts
export class Doctor_Master {

  DoctorID: number;
  CMPID: number;
  // DoctorName: string = "";
  Address1 = '';
  Address2: string = null;
  Address3: string = null;
  LocationID: number;
  Designation: string = null;
  RegistrationNumber: string = null;
  EngagementType: number;
  Phone1 = '';
  Phone2: string = null;
  EmailID: string = null;
  Title: string;
  FirstName: string;
  LastName: string;
  MiddleName: string;
  Gender: number;
  DateofBirth: Date;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  IsActive: boolean;
}